feat: separate withdrawal/deposit
