+++ tests/src/lib.rs
@@ -1,4 +1,5 @@
 pub mod config;
-pub mod deposit_withdrawal;
+pub mod deposit;
 pub mod send;
 pub mod utils;
+pub mod withdrawal;
