+++ tests/src/withdrawal.rs
@@ -1,6 +1,10 @@
+use crate::{
+    config::TestConfig,
+    send::send_transfers,
+    utils::{get_balance_on_intmax, get_block_builder_url},
+};
 use alloy::{primitives::B256, providers::Provider as _};
 use anyhow::Context;
-use intmax2_cli::cli::deposit::fetch_predicate_permission;
 use intmax2_client_sdk::{
     client::{
         client::Client, fee_payment::generate_fee_payment_memo,
@@ -14,129 +18,12 @@ use intmax2_client_sdk::{
         utils::time::sleep_for,
     },
 };
-use intmax2_interfaces::{
-    api::withdrawal_server::interface::WithdrawalStatus, data::deposit_data::TokenType,
-};
+use intmax2_interfaces::api::withdrawal_server::interface::WithdrawalStatus;
 use intmax2_zkp::{
     common::{salt::Salt, transfer::Transfer, withdrawal::get_withdrawal_nullifier},
-    ethereum_types::{address::Address, u256::U256, u32limb_trait::U32LimbTrait},
+    ethereum_types::{u256::U256, u32limb_trait::U32LimbTrait},
 };
 use std::time::Duration;
-
-use crate::{
-    config::TestConfig,
-    send::send_transfers,
-    utils::{calculate_balance_with_gas_deduction, get_balance_on_intmax, get_block_builder_url},
-};
-
-pub async fn single_deposit(
-    config: &TestConfig,
-    client: &Client,
-    eth_private_key: B256,
-) -> anyhow::Result<()> {
-    let key = generate_intmax_account_from_eth_key(eth_private_key);
-    let depositor = get_address_from_private_key(eth_private_key);
-    let gas_limit = 200000;
-    let deposit_amount = calculate_balance_with_gas_deduction(
-        &client.liquidity_contract.provider,
-        depositor,
-        2,
-        gas_limit,
-    )
-    .await?;
-
-    let depositor = convert_address_to_intmax(depositor);
-    let deposit_result = client
-        .prepare_deposit(
-            depositor,
-            key.pubkey,
-            convert_u256_to_intmax(deposit_amount),
-            TokenType::NATIVE,
-            Address::default(),
-            0.into(),
-            false,
-        )
-        .await?;
-
-    let deposit_data = deposit_result.deposit_data.clone();
-    let aml_permission = fetch_predicate_permission(
-        client,
-        depositor,
-        deposit_data.pubkey_salt_hash,
-        deposit_data.token_type,
-        deposit_data.amount,
-        deposit_data.token_address,
-        deposit_data.token_id,
-    )
-    .await?;
-    let eligibility_permission = vec![];
-
-    client
-        .liquidity_contract
-        .deposit_native(
-            eth_private_key,
-            None,
-            deposit_data.pubkey_salt_hash,
-            deposit_data.amount,
-            &aml_permission,
-            &eligibility_permission,
-        )
-        .await?;
-
-    // Wait for the deposit to be synced to the validity prover
-    let mut retries = 0;
-    loop {
-        if retries >= config.deposit_sync_check_retries {
-            return Err(anyhow::anyhow!("Deposit is not synced to validity prover"));
-        }
-        let deposit_info = client
-            .validity_prover
-            .get_deposit_info(deposit_data.pubkey_salt_hash)
-            .await?;
-        if deposit_info.is_some() {
-            break;
-        }
-        log::warn!("Deposit is not synced to validity prover, retrying...");
-        tokio::time::sleep(Duration::from_secs(config.deposit_sync_check_interval)).await;
-        retries += 1;
-    }
-    log::info!("Deposit is synced to validity prover");
-
-    // Wait for the deposit to be relayed to the L2
-    let mut retries = 0;
-    loop {
-        if retries >= config.deposit_relay_check_retries {
-            return Err(anyhow::anyhow!("Deposit is not synced to validity prover"));
-        }
-        let deposit_info = client
-            .validity_prover
-            .get_deposit_info(deposit_data.pubkey_salt_hash)
-            .await?
-            .ok_or(anyhow::anyhow!(
-                "Deposit is disappeared from validity prover"
-            ))?;
-        if deposit_info.block_number.is_some() {
-            break;
-        }
-        log::warn!("Deposit is not synced to L2, retrying...");
-        tokio::time::sleep(Duration::from_secs(config.deposit_relay_check_interval)).await;
-        retries += 1;
-    }
-    log::info!("Deposit is relayed to L2");
-
-    // sync balance
-    client.sync(key).await?;
-    log::info!("Synced balance");
-
-    let intmax_balance = get_balance_on_intmax(client, key).await?;
-    if intmax_balance < deposit_data.amount {
-        return Err(anyhow::anyhow!(
-            "Deposit is not reflected in the balance: {}",
-            intmax_balance
-        ));
-    }
-    Ok(())
-}
 
 pub async fn single_withdrawal(
     config: &TestConfig,
